Tabs working

+++ .eslintrc.js
@@ -42,7 +42,7 @@ module.exports = {
         "no-undef": "off",
         "no-nested-ternary":"off",
         "no-unused-expressions":"off",
-		"quotes": [2, "single", { "avoidEscape": true }]
+		"quotes": [2, "double", { "avoidEscape": true }]
 
     },
     "globals": {
